outbound: Differentiate generic TCP utils from opaque proxying.
