+++ linkerd/app/outbound/src/lib.rs
@@ -11,6 +11,7 @@ pub mod http;
 mod ingress;
 pub mod logical;
 mod metrics;
+pub mod opaq;
 mod switch_logical;
 pub mod tcp;
 #[cfg(test)]
