+++ linkerd/app/outbound/src/tcp/connect.rs
@@ -35,34 +35,6 @@ impl Outbound<()> {
     }
 }
 
-impl<C> Outbound<C> {
-    pub fn push_tcp_forward<T, I>(
-        self,
-    ) -> Outbound<
-        svc::ArcNewService<
-            T,
-            impl svc::Service<I, Response = (), Error = EndpointError, Future = impl Send> + Clone,
-        >,
-    >
-    where
-        T: svc::Param<Remote<ServerAddr>> + Clone + Send + Sync + 'static,
-        I: io::AsyncRead + io::AsyncWrite + io::PeerAddr + std::fmt::Debug + Send + Unpin + 'static,
-        C: svc::MakeConnection<T> + Clone + Send + Sync + 'static,
-        C::Connection: Send + Unpin,
-        C::Metadata: Send + Unpin,
-        C::Future: Send,
-    {
-        self.map_stack(|_, _, conn| {
-            conn.push(svc::stack::WithoutConnectionMetadata::layer())
-                .push_new_thunk()
-                .push_on_service(super::Forward::layer())
-                .push(svc::NewMapErr::layer_from_target())
-                .push(svc::ArcNewService::layer())
-                .check_new_service::<T, I>()
-        })
-    }
-}
-
 // === impl PreventLoopback ===
 
 impl<S> PreventLoopback<S> {
@@ -134,59 +106,5 @@ where
             addr: target.param(),
             source,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        svc::{self, NewService, ServiceExt},
-        test_util::*,
-        transport::{ClientAddr, Local},
-    };
-    use std::net::SocketAddr;
-
-    #[derive(Clone)]
-    struct Target(SocketAddr);
-
-    impl svc::Param<Option<crate::http::Version>> for Target {
-        fn param(&self) -> Option<crate::http::Version> {
-            None
-        }
-    }
-
-    impl svc::Param<Remote<ServerAddr>> for Target {
-        fn param(&self) -> Remote<ServerAddr> {
-            Remote(ServerAddr(self.0))
-        }
-    }
-
-    #[tokio::test]
-    async fn forward() {
-        let _trace = linkerd_tracing::test::trace_init();
-
-        let addr = SocketAddr::new([192, 0, 2, 2].into(), 2222);
-        let (rt, _shutdown) = runtime();
-        let stack = Outbound::new(default_config(), rt)
-            .with_stack(svc::mk(move |Target(a): Target| {
-                assert_eq!(a, addr);
-                let mut io = support::io();
-                io.write(b"hello").read(b"world");
-                future::ok::<_, support::io::Error>((
-                    io.build(),
-                    Local(ClientAddr(([0, 0, 0, 0], 0).into())),
-                ))
-            }))
-            .push_tcp_forward()
-            .into_inner();
-
-        let mut io = support::io();
-        io.read(b"hello").write(b"world");
-        stack
-            .new_service(Target(addr))
-            .oneshot(io.build())
-            .await
-            .expect("forward must complete successfully");
     }
 }
